refactor(infoWidget): info widget uses a template
